+++ data/outport/grpcadapter/errors.go
@@ -1,4 +1,4 @@
-package grpc
+package grpcadapter
 
 import "errors"
