+++ data/outport/grpcadapter/grpcServer_test.go
@@ -1,4 +1,4 @@
-package grpc
+package grpcadapter
 
 import (
 	"context"
