+++ data/outport/grpcadapter/interface.go
@@ -1,4 +1,4 @@
-package grpc
+package grpcadapter
 
 import (
 	"context"
